+++ src/Molecules/AccordionItem/AccordionItem.tsx
@@ -6,14 +6,14 @@ import { AccordionItemComponent } from './AccordionItem.types';
 
 const Item = styled.div`
   & + & {
-    border-top: 1px solid ${p => p.theme.color.divider};
+    border-top: 1px solid ${(p) => p.theme.color.divider};
   }
 `;
 
 const Panel = styled.div`
-  padding-top: ${p => p.theme.spacing.unit(1)}px;
-  padding-bottom: ${p => p.theme.spacing.unit(3)}px;
-  padding-left: ${p => p.theme.spacing.unit(6)}px;
+  padding-top: ${(p) => p.theme.spacing.unit(1)}px;
+  padding-bottom: ${(p) => p.theme.spacing.unit(3)}px;
+  padding-left: ${(p) => p.theme.spacing.unit(6)}px;
 
   &[hidden] {
     display: none;
@@ -27,16 +27,16 @@ const Button = styled.button`
   display: flex;
   width: 100%;
   border: 0;
-  padding-top: ${p => p.theme.spacing.unit(3)}px;
+  padding-top: ${(p) => p.theme.spacing.unit(3)}px;
   padding-right: 0;
-  padding-bottom: ${p => p.theme.spacing.unit(3)}px;
-  padding-left: ${p => p.theme.spacing.unit(6)}px;
+  padding-bottom: ${(p) => p.theme.spacing.unit(3)}px;
+  padding-left: ${(p) => p.theme.spacing.unit(6)}px;
   background-color: transparent;
   cursor: pointer;
 
   &:hover,
   &:focus {
-    background-color: ${p => p.theme.color.background};
+    background-color: ${(p) => p.theme.color.background};
   }
 `;
 
@@ -64,7 +64,7 @@ export const AccordionItem: AccordionItemComponent = React.forwardRef(
     const isControlled = isBoolean(controlledExpand);
     const expanded = isControlled ? controlledExpand : expandedInternal;
 
-    const clickHandler: React.MouseEventHandler = e => {
+    const clickHandler: React.MouseEventHandler = (e) => {
       if (!isControlled) {
         setExpandedInternal(!expandedInternal);
       }
@@ -80,13 +80,13 @@ export const AccordionItem: AccordionItemComponent = React.forwardRef(
 
     return (
       <Item className={className}>
-        <Typography as={as} type="secondary">
+        <Typography as={as} type="secondary" weight="bold">
           <Button type="button" aria-expanded={expanded} onClick={clickHandler} ref={ref}>
             <IconWrapper>
               {expanded ? (
-                <Icon.Minus size={3} fill={t => t.color.cta} />
+                <Icon.Minus size={3} fill={(t) => t.color.cta} />
               ) : (
-                <Icon.Plus size={3} fill={t => t.color.cta} />
+                <Icon.Plus size={3} fill={(t) => t.color.cta} />
               )}
             </IconWrapper>
             {title}