+++ src/Molecules/Drawer/Drawer.types.ts
@@ -1,9 +1,12 @@
+type HtmlDivProps = React.HTMLAttributes<HTMLDivElement>;
+
 export type Props = {
   className?: string;
   /** @default true */
   closeOnClickOutside?: boolean;
   /** Useful when integrating with FadedScroll component */
   disableContentStyle?: boolean;
+  footer?: React.ReactNode;
   onClose?: Function;
   title?: React.ReactNode;
   open?: boolean;
@@ -13,7 +16,7 @@ export type Props = {
   onAnimationComplete?: () => void;
   disableInitialAnimation?: boolean;
   preventOnClickOutsideDataAttributes?: string[];
-};
+} & Omit<HtmlDivProps, 'title'>;
 
 export type TitleProps = {
   title: React.ReactNode;