+++ src/Molecules/Input/Select/lib/utils.ts
@@ -1,0 +1,11 @@
+export const getSingleSelectValue = (value: any[]): any[] => {
+  if (!value.length) {
+    return [];
+  }
+
+  if (!value.some((x) => x.options)) {
+    return value;
+  }
+
+  return value.find((x) => x.options).options;
+};