+++ src/Molecules/Switch/Switch.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import { Typography, Flexbox, VisuallyHidden } from '../..';
-import { isElement } from '../../common/utils';
+import { isBoolean, isElement } from '../../common/utils';
 import NormalizedElements from '../../common/NormalizedElements';
 import { Props } from './Switch.types';
 
@@ -16,26 +16,26 @@ const Label = styled.label`
 const Knob = styled.span`
   background: white;
   display: block;
-  height: ${p => p.theme.spacing.unit(KNOB_SIZE)}px;
-  width: ${p => p.theme.spacing.unit(KNOB_SIZE)}px;
+  height: ${(p) => p.theme.spacing.unit(KNOB_SIZE)}px;
+  width: ${(p) => p.theme.spacing.unit(KNOB_SIZE)}px;
   position: absolute;
   top: 50%;
   left: 0;
-  border: 1px solid ${p => p.theme.color.inputBorder};
-  margin-top: -${p => p.theme.spacing.unit(KNOB_SIZE / 2)}px;
-  border-radius: ${p => p.theme.spacing.unit(KNOB_SIZE / 2)}px;
+  border: 1px solid ${(p) => p.theme.color.inputBorder};
+  margin-top: -${(p) => p.theme.spacing.unit(KNOB_SIZE / 2)}px;
+  border-radius: ${(p) => p.theme.spacing.unit(KNOB_SIZE / 2)}px;
   box-sizing: border-box;
-  box-shadow: 0px 1px 3px 1px ${p => p.theme.color.shadowSwitch};
+  box-shadow: 0px 1px 3px 1px ${(p) => p.theme.color.shadowSwitch};
   transition: transform 0.2s cubic-bezier(0.18, 0.9, 0.35, 1.15);
 `;
 
 const Track = styled.span`
   display: block;
-  height: ${p => p.theme.spacing.unit(TRACK_HEIGHT)}px;
-  width: ${p => p.theme.spacing.unit(TRACK_WIDTH)}px;
-  margin: ${p => p.theme.spacing.unit((KNOB_SIZE - TRACK_HEIGHT) / 2)}px 0;
-  background-color: ${p => p.theme.color.inputBorder};
-  border-radius: ${p => p.theme.spacing.unit(TRACK_HEIGHT / 2)}px;
+  height: ${(p) => p.theme.spacing.unit(TRACK_HEIGHT)}px;
+  width: ${(p) => p.theme.spacing.unit(TRACK_WIDTH)}px;
+  margin: ${(p) => p.theme.spacing.unit((KNOB_SIZE - TRACK_HEIGHT) / 2)}px 0;
+  background-color: ${(p) => p.theme.color.inputBorder};
+  border-radius: ${(p) => p.theme.spacing.unit(TRACK_HEIGHT / 2)}px;
   transition: background-color 0.2s ease-out;
 `;
 
@@ -51,21 +51,21 @@ const Button = styled(NormalizedElements.Button)`
 
   &[aria-checked='true'] {
     ${Track} {
-      background-color: ${p => p.theme.color.cta};
+      background-color: ${(p) => p.theme.color.cta};
     }
 
     ${Knob} {
-      transform: translate(${p => p.theme.spacing.unit(TRACK_WIDTH - KNOB_SIZE)}px);
+      transform: translate(${(p) => p.theme.spacing.unit(TRACK_WIDTH - KNOB_SIZE)}px);
     }
   }
 
   &[disabled] {
     ${Track} {
-      background-color: ${p => p.theme.color.disabledBackground};
+      background-color: ${(p) => p.theme.color.disabledBackground};
     }
 
     ${Knob} {
-      border: 1px solid ${p => p.theme.color.disabledBackground};
+      border: 1px solid ${(p) => p.theme.color.disabledBackground};
     }
   }
 `;
@@ -78,17 +78,23 @@ export const Switch: React.FC<Props> = ({
   onClick,
   checkedInitially = false,
   checked: checkedControlled,
+  readOnly,
 }) => {
-  const isControlled = typeof checkedControlled !== 'undefined';
+  const isControlled = isBoolean(checkedControlled);
   const titleNode = isElement(label) ? label : <Typography type="secondary">{label}</Typography>;
-  const [checked, setChecked] = useState(checkedInitially);
-  const clickHandler = (e: React.MouseEvent) => {
-    const nextState = !checked;
+  const [checkedInternal, setCheckedInternal] = useState(checkedInitially);
+  const checked = isControlled ? checkedControlled : checkedInternal;
 
-    setChecked(nextState);
+  const internalClickHandler = (e: React.MouseEvent) => {
+    const nextCheckedState = !checked;
+
+    if (!readOnly) {
+      setCheckedInternal(nextCheckedState);
+    }
 
     if (onClick) {
-      setTimeout(() => onClick(e, nextState));
+      // setTimout from previous commit for animation reasons
+      setTimeout(() => onClick(e, readOnly ? checked : nextCheckedState));
     }
   };
 
@@ -103,9 +109,10 @@ export const Switch: React.FC<Props> = ({
             className={className}
             type="button"
             role="switch"
-            aria-checked={isControlled ? checkedControlled : checked}
-            onClick={isControlled ? onClick : clickHandler}
+            aria-checked={checked}
+            onClick={isControlled ? onClick : internalClickHandler}
             disabled={disabled}
+            aria-readonly={readOnly}
           >
             <ButtonContent>
               <Knob />