+++ src/Molecules/Switch/Switch.types.ts
@@ -9,5 +9,9 @@ export type Props = {
   disabled?: boolean;
   label: string | React.ReactNode;
   hiddenLabel?: boolean;
-  onClick?: (e: React.MouseEvent, checked: boolean) => void;
+  /**
+   * In controlled mode the parameter checked is unnecessary and therefore dropped
+   */
+  onClick?: (e: React.MouseEvent, checked?: boolean) => void;
+  readOnly?: boolean;
 };